optimizer/requirements.txt
# Core Framework
fastapi>=0.115.12
uvicorn[standard]>=0.34.2
gunicorn>=23.0.0
python-dotenv>=1.1.0

# Database - Optimized for Supabase Transaction Pooler
SQLAlchemy>=2.0.41
psycopg2-binary>=2.9.10
alembic>=1.16.2

# Supabase SDK
supabase>=2.15.3

# Data Validation
pydantic>=2.11.7
pydantic[email]
pydantic-settings>=2.0.3

# Authentication
python-jose[cryptography]

# AI/ML & Document Processing
baml-py>=0.212.0
markitdown[all]

# File Handling
python-multipart>=0.0.20

# Financial Data & Analysis
yfinance>=0.2.63
fredapi>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
riskfolio-lib>=7.0.1
matplotlib>=3.8.0
cvxpy>=1.5.2
clarabel>=0.6.0
statsmodels>=0.13.5
arch>=7.0
xlsxwriter>=3.1.2
networkx>=3.0
astropy>=5.1
pybind11>=2.10.1

# HTTP & Web Scraping
httpx>=0.28.1
aiohttp>=3.11.13
beautifulsoup4>=4.12.3
lxml>=5.3.0
requests[socks]>=2.32.0  # SOCKS5 proxy support for PIA
pysocks>=1.7.1  # SOCKS proxy support

# Monitoring & System
psutil>=7.1.0
prometheus-client>=0.21.1

# Testing
pytest>=8.4.1
pytest-asyncio>=1.1.0

# Progress Bars
tqdm>=4.67.1